handle resource variables
